Update fetch message

+++ lib/domain/impl/fetch_message_impl.dart
@@ -1,5 +1,7 @@
 import 'dart:async';
 
+import 'package:either_dart/either.dart';
+import 'package:vita_client_app/data/model/entity/message.dart';
 import 'package:vita_client_app/domain/fetch_message.dart';
 import 'package:vita_client_app/repository/message_repository.dart';
 
@@ -9,13 +11,15 @@ class FetchMessageImpl implements FetchMessage {
   FetchMessageImpl(this._repository);
 
   @override
-  Future<void> call() async {
+  Future<Either<Error, List<Message>>> call() async {
     var response = await _repository.getMessage();
-    if (response.isSuccessful) {
-      if (response.body != null) {
-        await _repository.deleteMessage();
-        _repository.inserts(response.body!);
-      }
+    if (response.isSuccessful && response.body != null) {
+      var messages = response.body!;
+      await _repository.deleteMessage();
+      _repository.inserts(messages);
+      return Right(messages);
+    } else {
+      return Left(response.error! as Error);
     }
   }
 }
